backend/pyproject.toml
[tool.poetry]
name = "web-app-template"
version = "0.1.0"
description = ""
authors = ["Mohammed Fuad <mohammedfuadalnashriy@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
django = "^5.1.3"
psycopg2-binary = "^2.9.10"
dj-database-url = "^2.3.0"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
dj-rest-auth = { extras = ["with-social"], version = "^7.0.0" }
django-cors-headers = "^4.6.0"
djangorestframework-simplejwt = "^5.3.1"
ollama = "^0.4.7"
channels = {extras = ["daphne"], version = "^4.2.0"}
crawl4ai = "^0.4.247"
beautifulsoup4 = "^4.12.3"
sumy = "^0.11.0"
markdown = "^3.7"
pypdf2 = "^3.0.1"
reportlab = "^4.2.5"
weasyprint = "^64.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.4"
pre-commit = "^4.0.1"


[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "migrations",
]

# Same as Black.
line-length = 120
indent-width = 4

# Assume Python 3.10
target-version = "py310"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
select = ["E", "F", "B", "I"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
